requirements.txt
# Core dependencies
typer>=0.9.0
pyyaml>=6.0.0
jsonschema>=4.0.0

# Vector DB (for Phase 2)
# chromadb>=0.4.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
